requirements.txt
flask
flasgger
gunicorn
line-bot-sdk
rich